requirements.txt
requests~=2.28.2
requests-cache~=1.0.0
python-dotenv~=1.0.0
pyTelegramBotAPI~=4.10.0
Flask~=2.3.2